+++ db/migrate/20230213132720_create_recipe_food.rb
@@ -1,7 +1,7 @@
 class CreateRecipeFood < ActiveRecord::Migration[7.0]
   def change
     create_table :recipe_foods do |t|
-      t.decimal :quantity
+      t.float :quantity
       t.references :recipe, null: false, foreign_key: true
       t.references :food, null: false, foreign_key: true
